Build docker image with database
During the current phase of development, this is a lot easier
than setting up a volume and manually syncing the file. I'm not
sure if it's a good idea in the long run, as we may want to
update the database more often than the docker image, but let's
run with it until it becomes a problem.