Replace bar chart with heat maps
The bar chart doesn't really work as a component. The word stats
are too sparse and the component becomes too wide without side
scrolling, which is poor UX.

Hope this heat map works better. Different component in desktop
and mobile mode to accomodate for the different modes.

Refinements will follow.